requirements.txt
absl-py==0.2.2
astor==0.6.2
beautifulsoup4==4.6.0
bleach==1.5.0
certifi==2018.4.16
chardet==3.0.4
desire2download==1.0.0
Django==2.0.7
dropbox==9.0.0
gast==0.2.0
grpcio==1.12.0
gunicorn==19.9.0
html5lib==0.9999999
idna==2.6
lxml==4.2.1
Markdown==2.6.11
MechanicalSoup==0.10.0
numpy==1.14.3
Pillow==5.1.0
pipenv==2018.7.1
protobuf==3.5.2.post1
python-dateutil==2.7.3
pytz==2018.5
requests==2.18.4
selenium==3.12.0
six==1.11.0
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
torch==0.4.0
torchvision==0.2.1
urllib3==1.22
virtualenv==16.0.0
virtualenv-clone==0.3.0
Werkzeug==0.14.1
whitenoise==3.3.1
